created conftest.py and pytest.ini, added parametrize testcase and modified library.remove_book to return instead of print
