+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+markers =
+    P0: Priority 0 tests (must-run smoke tests)
+    P1: Priority 1 tests (high-priority regression tests)
+    P2: Priority 2 tests (medium-priority tests)
